modules/sensor_filter/requirements.txt
aiohttp==3.8.1
asyncssh==2.11.0
asyncio==3.4.3
azure-iot-device==2.12.0
iot-edge-logger==1.0.6
